Task-4/Pdf_Quiz_agent/requirements.txt
streamlit
PyPDF2
openai-agents[litellm]
pydantic
python-dotenv